Seperating Entity required tests
