+++ test/various.test.js
@@ -12,78 +12,6 @@ var describe = lab.describe
 var it = lab.it
 
 describe('Various misc', function () {
-  it('uses correct tx$ properties on entity actions for "transported" entities', function (done) {
-    var seneca1 = CreateInstance()
-    .ready(function () {
-      seneca1.add({cmd: 'test'}, function (args, cb) {
-        args.entity.save$(function (err, entitySaveResponse) {
-          if (err) return cb(err)
-
-          this.act({cmd: 'test2'}, function (err, test2Result) {
-            if (err) {
-              return cb(err)
-            }
-
-            cb(null, {
-              entity: entitySaveResponse.entity,
-              txBeforeEntityAction: args.tx$,
-              txInsideEntityAction: entitySaveResponse.tx,
-              txAfterEntityAction: test2Result.tx
-            })
-          })
-        })
-      })
-      .add({role: 'entity', cmd: 'save'}, function (args, cb) {
-        cb(null, { entity: args.ent, tx: args.tx$ })
-      })
-      .add({cmd: 'test2'}, function (args, cb) {
-        cb(null, { tx: args.tx$ })
-      })
-      .listen({type: 'tcp', port: 20103})
-
-      var seneca2 = CreateInstance()
-      .ready(function () {
-        seneca2.client({type: 'tcp', port: 20103})
-        this.act({cmd: 'test', entity: this.make$('test').data$({name: 'bar'})}, function (err, res) {
-          Assert(!err)
-
-          Assert(res.entity.name === 'bar')
-          Assert(res.txBeforeEntityAction === res.txInsideEntityAction)
-          Assert(res.txBeforeEntityAction === res.txAfterEntityAction)
-          done()
-        })
-      })
-    })
-  })
-
-  it('uses correct tx$ properties on entity actions for "non-transported" requests', function (done) {
-    CreateInstance()
-      .add({ cmd: 'test' }, function (args, cb) {
-        this.act({ cmd: 'test2' }, function (err, test2Result) {
-          if (err) {
-            return cb(err)
-          }
-
-          cb(null, { txBeforeEntityAction: args.tx$, txAfterEntityAction: test2Result.tx })
-        })
-      })
-      .add({ cmd: 'test2' }, function (args, cb) {
-        cb(null, { tx: args.tx$ })
-      })
-      .listen({ type: 'tcp', port: 20104 })
-      .ready(function () {
-        CreateInstance()
-          .client({ type: 'tcp', port: 20104 })
-          .ready(function () {
-            this.act({ cmd: 'test' }, function (err, res) {
-              Assert(!err)
-              Assert(res.txBeforeEntityAction === res.txAfterEntityAction)
-              done()
-            })
-          })
-      })
-  })
-
   // NOTE: SENECA_LOG=all will break this test as it counts log entries
   it('own-message', function (fin) {
     // a -> b -> a
